--- Catalog/Catalog 2.xcdatamodeld/Catalog.xcdatamodel.folder/contents
@@ -1,16 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
-<model type="com.apple.IDECoreDataModeler.DataModel" documentVersion="1.0" lastSavedToolsVersion="14135" systemVersion="17G65" minimumToolsVersion="Automatic" sourceLanguage="Swift" userDefinedModelVersionIdentifier="">
-    <entity name="CustomerMO" representedClassName="CustomerMO" syncable="YES" codeGenerationType="class">
-        <attribute name="appVersion" attributeType="String" syncable="YES"/>
-        <attribute name="dateOfVisit" attributeType="String" syncable="YES"/>
-        <attribute name="dateOfWedding" optional="YES" attributeType="String" syncable="YES"/>
-        <attribute name="dressesNames" attributeType="String" customClassName="NSArray" syncable="YES"/>
-        <attribute name="name" optional="YES" attributeType="String" syncable="YES"/>
-        <attribute name="region" optional="YES" attributeType="String" syncable="YES"/>
-        <attribute name="shopId" attributeType="String" syncable="YES"/>
-        <attribute name="surname" optional="YES" attributeType="String" syncable="YES"/>
-    </entity>
-    <elements>
-        <element name="CustomerMO" positionX="-54" positionY="-9" width="128" height="165"/>
-    </elements>
-</model>
