+++ Catalog/Utils/APIConnector.swift
