+++ Catalog/Utils/CoreDataManager.swift
