+++ Catalog/Utils/ShopIdManager.swift
