+++ Catalog/Utils/Validator.swift
