+++ Catalog/Utils/fonts/RobotoCondensed-Italic.ttf
